+++ changelog.md
@@ -5,6 +5,10 @@ Initial release
 # TODO
 
 * Test sleeping and multiplayer
+* Render types
+* Mixed textures and tint=true custom block model renderer
+* Effects reset to 0
+* Finish sprites menu
 
 # Roadmap 1.0.0